+++ R/getters_and_setters.R
@@ -70,7 +70,7 @@ i_match_ij_spectra = function(x, i = NULL, j = NULL, allow_negative = FALSE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' head(names(spec), n = 3)
 #' # by name
 #' spec1 = spec[ "species_7" , ]
@@ -117,7 +117,7 @@ i_match_ij_spectra = function(x, i = NULL, j = NULL, allow_negative = FALSE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec[ , 400:500] = spec[ , 400:500] * 1.2
 #' spec
 `[<-.spectra` = function(x, i, j, value){
@@ -159,7 +159,7 @@ i_match_ij_spectra = function(x, i = NULL, j = NULL, allow_negative = FALSE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' is.matrix(reflectance(spec))
 reflectance = function(x){
     UseMethod("reflectance")
@@ -178,7 +178,7 @@ reflectance = function(x){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' # scale all refletance values by 2
 #' reflectance(spec) = reflectance(spec) * 2
 `reflectance<-` = function(x, value){
@@ -215,7 +215,7 @@ reflectance.spectra = function(x){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' # is reflectance being forced to be between 0 and 1?
 #' enforce01(spec)
 enforce01 = function(x){
@@ -235,7 +235,7 @@ enforce01 = function(x){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #'
 #' # Force to be between 0 and 1.
 #' enforce01(spec) = TRUE
@@ -285,7 +285,7 @@ enforce01.spectra = function(x){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' names(spec)
 names.spectra = function(x){
     x$names
@@ -309,7 +309,7 @@ names.spectra = function(x){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' names(spec) = toupper(names(spec))
 `names<-.spectra` = function(x, value){
     x$names = i_names(value, nrow(x), prefix = NULL)
@@ -336,7 +336,7 @@ names.spectra = function(x){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' head(wavelengths(spec))
 wavelengths = function(x, min = NULL, max = NULL, return_num = TRUE){
     UseMethod("wavelengths")
@@ -357,7 +357,7 @@ wavelengths = function(x, min = NULL, max = NULL, return_num = TRUE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' wavelengths(spec) = wavelengths(spec) / 1000
 `wavelengths<-` = function(x, unsafe = FALSE, value){
     UseMethod("wavelengths<-")
@@ -418,7 +418,7 @@ wavelengths.spectra = function(x, min = NULL, max = NULL, return_num = TRUE) {
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec = normalize(spec)
 #' meta(spec, "normalization_magnitude")
 meta = function(x, label, sample, simplify = FALSE, quiet = TRUE){
@@ -440,7 +440,7 @@ meta = function(x, label, sample, simplify = FALSE, quiet = TRUE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' meta(spec, "random") = rnorm(nrow(spec), mean(10), sd = 2)
 `meta<-` = function(x, label, sample, value){
     UseMethod("meta<-")