+++ R/math_operator_overloads.R
@@ -10,7 +10,7 @@
 #'
 #' @examples
 #' library(spectrolab)
-#' s = as.spectra(spec_matrix_example)
+#' s = as.spectra(spec_matrix_example, name_idx = 1)
 #'
 #' # This will throw an error
 #' \dontrun{
@@ -37,7 +37,7 @@ t.spectra = function(x){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec  = as.spectra(spec_matrix_example)
+#' spec  = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec1 = spec * 2
 #' spec2 = spec + spec
 #' all(spec1 == spec2)