+++ R/stats_function_overloads.R
@@ -17,7 +17,7 @@
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' mean(spec)
 mean.spectra = function(x, na.rm = TRUE, keep_txt_meta = TRUE, ...){
     apply_by_band(x, base::mean, na.rm = na.rm, keep_txt_meta = keep_txt_meta, ...)
@@ -40,7 +40,7 @@ mean.spectra = function(x, na.rm = TRUE, keep_txt_meta = TRUE, ...){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' median(spec)
 median.spectra = function(x, na.rm = TRUE, keep_txt_meta = TRUE, ...){
     apply_by_band(x, stats::median, na.rm = na.rm, keep_txt_meta = keep_txt_meta, ...)
@@ -103,7 +103,7 @@ var.default = stats::var
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' var(spec)
 var.spectra = function(x, y = NULL, na.rm = TRUE, use){
     apply_by_band(x, stats::var, na.rm = na.rm, keep_txt_meta = TRUE)
@@ -145,7 +145,7 @@ sd.default = stats::sd
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' sd(spec)
 sd.spectra = function(x, na.rm = TRUE){
     apply_by_band(x, stats::sd, na.rm = na.rm, keep_txt_meta = TRUE)
@@ -171,7 +171,7 @@ sd.spectra = function(x, na.rm = TRUE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' quantile(spec, probs = c(0.25, 0.75))
 quantile.spectra = function(x,
                             probs = c(0.025, 0.25, 0.5, 0.75, 0.975),