+++ R/transform_spectra.R
@@ -26,7 +26,7 @@
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec_mean = apply_by_band(spec, mean)
 apply_by_band = function(x, fun, na.rm = TRUE, keep_txt_meta = TRUE, name = NULL, ...){
     UseMethod("apply_by_band")
@@ -103,7 +103,7 @@ apply_by_band.spectra = function(x, fun, na.rm = TRUE, keep_txt_meta = TRUE, nam
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec_mean = aggregate(spec, by = names(spec), mean, try_keep_txt(mean))
 aggregate.spectra = function(x, by, FUN, FUN_meta = NULL, ...){
 
@@ -144,9 +144,9 @@ aggregate.spectra = function(x, by, FUN, FUN_meta = NULL, ...){
 #' library(spectrolab)
 #'
 #' # Create dummy spectra datasets. Pretend that these are all different...
-#' s1 = as.spectra(spec_matrix_example)
-#' s2 = as.spectra(spec_matrix_example)
-#' s3 = as.spectra(spec_matrix_example)
+#' s1 = as.spectra(spec_matrix_example, name_idx = 1)
+#' s2 = as.spectra(spec_matrix_example, name_idx = 1)
+#' s3 = as.spectra(spec_matrix_example, name_idx = 1)
 #'
 #' # combine 2 spectra objects
 #' s_1and2 = combine(s1, s2)
@@ -217,7 +217,7 @@ combine.spectra = function(s1, s2){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec_list = split(spec, names(spec))
 split.spectra = function(x, f, drop = FALSE, ...){
 
@@ -260,7 +260,7 @@ split.spectra = function(x, f, drop = FALSE, ...){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #'
 #' # remove spec of species with less than 4 samples
 #' spec = subset_by(spec, by = names(spec), n_min = 4, n_max = Inf)
@@ -352,7 +352,7 @@ subset_by.spectra = function(x, by, n_min, n_max, random = TRUE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec = normalize(spec)
 normalize = function(x, quiet = FALSE, ...){
     UseMethod("normalize")
@@ -517,7 +517,7 @@ i_smooth_mav_spectra = function(x, n = NULL, save_wvls_to_meta = TRUE){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec = smooth(spec, parallel = FALSE)
 smooth = function(x, method = "spline", ...){
     UseMethod("smooth")
@@ -573,7 +573,7 @@ smooth.spectra = function(x, method = "spline", ...){
 #'
 #' @examples
 #' library(spectrolab)
-#' spec = as.spectra(spec_matrix_example)
+#' spec = as.spectra(spec_matrix_example, name_idx = 1)
 #' spec = resample(spec, new_wvls = seq(400, 2400, 0.5), parallel = FALSE)
 resample = function(x, new_wvls, ...) {
     UseMethod("resample")