+++ tests/testthat/test_combine.R
@@ -3,7 +3,7 @@ library("spectrolab")
 
 context("Spectra combine")
 
-spec = as.spectra(spec_matrix_example)
+spec = as.spectra(spec_matrix_example, name_idx = 1)
 s1   = spec
 wavelengths(s1) = seq(ncol(s1))
 s2   = spec[ , 400:2000]