+++ NEWS.md
@@ -1,3 +1,14 @@
+# spectrolab X.X.X (XXXX-XX-XX)
+
+## major 
+* *NOT backwards compatible!*. Now, as.spectra() does not assume by default that
+  the first column of a matrix or data.frame contains the names for the spectra.
+  Users must explicitly use the `name_idx` argument to pick the correct column
+
+## minor
+* XXX
+
+
 # spectrolab 0.0.5 (2018-05-31)
 
 ## major